split out cli from wflib
